Test against Py 3.11 + Break out non-DryRun-executing functions in `blackbox.py` into new script `inspector.py` (#48)
