+++ .github/workflows/build.yml
@@ -17,7 +17,7 @@ jobs:
     container: python:${{ matrix.python }}
     strategy:
       matrix:
-        python: [ "3.9", "3.10" ]
+        python: ["3.9", "3.10", "3.11"]
     steps:
       - run: python3 --version
       - name: Check out code
@@ -30,7 +30,7 @@ jobs:
     runs-on: ubuntu-20.04
     strategy:
       matrix:
-        python: [ "3.9", "3.10" ]
+        python: ["3.9", "3.10", "3.11"]
     steps:
       - name: Check out code
         uses: actions/checkout@v2
@@ -45,36 +45,13 @@ jobs:
           expected="${{ matrix.python }}"
           echo $installed
           [[ $installed =~ "Python ${expected}" ]] && echo "Configured Python" || (echo "Failed to configure Python" && exit 1)
-      - name: Install required os level applications
-        run: |
-          sudo apt update -y
-          sudo apt install -y curl git nodejs
-          sudo apt -y install ca-certificates curl gnupg lsb-release
-          sudo curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg
-          sudo echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://download.docker.com/linux/ubuntu \
-            $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
-          sudo apt update
-          sudo apt -y install docker-ce docker-ce-cli containerd.io
-      - name: Setup docker compose
-        run: |
-           sudo curl -L "https://github.com/docker/compose/releases/download/1.29.2/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
-           sudo chmod +x /usr/local/bin/docker-compose
-           sudo ln -s /usr/local/bin/docker-compose /usr/bin/docker-compose
-           docker-compose --version
-      - name: Check cache for Docker layers
-        uses: satackey/action-docker-layer-caching@v0.0.11
-        # Ignore the failure of a step and avoid terminating the job.
-        continue-on-error: true
-        with:
-          key: docker-layer-caching-${{ github.workflow }}-${{ hashFiles(env.SANDBOX_CONFIG, env.SANDBOX_GENESIS) }}-{hash}
-          restore-keys: docker-layer-caching-${{ github.workflow }}-${{ hashFiles(env.CONFIG, env.SANDBOX_GENESIS) }}-
-      - name: Create sandbox
-        uses: lucasvanmol/algorand-sandbox-action@v1
-        with:
-          config: ${{ env.SANDBOX_CONFIG }}
+      - name: Start algod
+        run: make algod-start
       - name: Setup integration test environment
         run: make pip-development unit-test
       - name: Setup Jupyter notebooks environment
         run: make pip-notebooks
-      - name: Run integration tests and Jupyter notebooks
+      - name: Run integration tests and Jupyter notebooks tests
         run: make gh-blackbox
+      - name: Stop algod
+        run: make algod-stop
