+++ Makefile
@@ -42,6 +42,15 @@ notebooks-test:
 
 all-tests: lint unit-test integration-test notebooks-test
 
+###### Algod / Docker #####
+
+algod-start:
+	docker compose up -d algod --wait
+
+algod-stop:
+	docker compose stop algod
+
+
 ###### Local Only ######
 
 # assumes installations of pipx, build and tox via:
@@ -60,7 +69,7 @@ local-blackbox: local-blackbox-smoke integration-test
 NOTEBOOK = notebooks/quadratic_factoring_game.ipynb
 # assumes already ran `make pip-notebooks`
 local-notebook:
-	 jupyter retro $(NOTEBOOK)
+	jupyter retro $(NOTEBOOK)
 
 # assumes act is installed, e.g. via `brew install act`:
 local-gh-simulate:
